refactor: move ocean_heat_content to submod
