+++ src/stactools/noaa_cdr/ocean_heat_content/asset-metadata.json
